requirements.txt
grpcio
grpcio-tools
git+https://github.com/yitech/coin-infra.git